helpdesk/requirements.txt
Django==6.0.2
django-environ==0.12.0
psycopg2-binary==2.9.11
python-magic==0.4.27
python-magic-bin==0.4.14